rename parser module to nonfungible_dot_com
from non_fungible_dot_com, for consistency
